Added integration tests based on test lang
